solve problem where stop are not considered by merging stop and station together

+++ phase1/src/AdminUser.java
@@ -2,7 +2,7 @@ class AdminUser {
 
     private String userName = "Admin";
     private String password;
-    static int totalFare;
+    static double totalFare;
     static int totalStations;
